+++ src/eez/modules/psu/dlog_view.cpp
@@ -239,7 +239,7 @@ void stateManagment() {
     }
 }
 
-float getValue(int rowIndex, int columnIndex, float *max) {
+float getValue(uint32_t rowIndex, uint8_t columnIndex, float *max) {
     uint32_t blockElementAddress = (rowIndex * getNumElementsPerRow() + columnIndex) * sizeof(BlockElement);
 
     uint32_t blockIndex = blockElementAddress / BLOCK_SIZE;
